+++ lib/algoliax/resources/index.ex
@@ -24,33 +24,71 @@ defmodule Algoliax.Resources.Index do
   end
 
   def replicas_names(module, settings, replica_index) do
-    settings
-    |> replicas_settings()
+    module
+    |> replicas_settings(settings)
     |> Enum.map(fn replica_settings ->
       index_name(module, replica_settings)
       |> Enum.at(replica_index)
     end)
   end
 
-  def replicas_settings(settings) do
-    replicas = Keyword.get(settings, :replicas, [])
-
-    Enum.map(replicas, fn replica ->
+  def replicas_settings(module, settings) do
+    settings
+    |> Keyword.get(:replicas, [])
+    |> Enum.filter(fn replica -> should_be_updated?(module, replica) end)
+    |> Enum.map(fn replica ->
       case Keyword.get(replica, :inherit, true) do
         true ->
           replica_algolia_settings = algolia_settings(replica)
-          primary_algolia_setttings = algolia_settings(settings)
+          primary_algolia_settings = algolia_settings(settings)
 
           Keyword.put(
             replica,
             :algolia,
-            Keyword.merge(primary_algolia_setttings, replica_algolia_settings)
+            Keyword.merge(primary_algolia_settings, replica_algolia_settings)
           )
 
         false ->
           replica
       end
     end)
+  end
+
+  defp should_be_updated?(module, replica) do
+    index_name = Keyword.get(replica, :index_name, nil)
+
+    error_message =
+      "`if` must be `nil|true|false` or be the name of a 0-arity func which returns a boolean."
+
+    value = Keyword.get(replica, :if, nil)
+
+    cond do
+      # No config, defaults to true
+      is_nil(value) ->
+        true
+
+      # Boolean, use this value
+      value == true || value == false ->
+        value
+
+      # Name of a 0-arity func
+      is_atom(value) ->
+        if module.__info__(:functions) |> Keyword.get(value) == 0 do
+          apply(module, value, []) == true
+        else
+          raise Algoliax.InvalidReplicaConfigurationError, %{
+            index_name: index_name,
+            error: error_message
+          }
+        end
+
+      # Any other value, raise an error
+      true ->
+        raise Algoliax.InvalidReplicaConfigurationError, %{
+          index_name: index_name,
+          error: error_message
+        }
+    end
   end
 
   def get_settings(module, settings) do
@@ -80,8 +118,8 @@ defmodule Algoliax.Resources.Index do
   end
 
   def configure_replicas(module, settings) do
-    settings
-    |> replicas_settings()
+    module
+    |> replicas_settings(settings)
     |> Enum.map(fn replica_settings ->
       configure_index(module, replica_settings)
     end)